Update patch set 2
Patch Set 2:

@Max,

I'm not 100% yet, I'm still having issues getting the dt2w to work, and I have yet to add dt2s.

The "knock-on" feature should just be a simple sysfs toggle, as the functionality is already included in the kernel. The device is already pre-configured for this.

Initially, I wanted to use a HAF like CM has it implemented with the cmhw, but that would involve too much. 

Thinking about what you had said earlier, this does make much more sense to be done at a device level, as it really only affects this specific device(s). Adding this in as an app, much more functionality can be added as well, exactly as you have done for oppo devices.

I'm currently working with this: https://github.com/jakew02/android_device_lge_g2-common/commits/tap_gestures

which is basically a copy of N1/Find5Parts (i will fix authorship before uploading here).  Just a little delayed as I'm working on completely re-doing the jflte repos, as the "unification" was half-assed and broken by those who did it.

Patch-set: 2